Updated to allow installation without the form interface
Updated the index to allow installation of plugins without using the
form interface. This is achieved by using the
$_SESSION['unified-installer-plugin-installation-values'] variable and
setting its members to the values needed by the installer. These values
include: database server, database name, database username and password,
as well as the prefix of the database tables. If this path is chosen,
every plugin found in the plugins directory will be installed, no option
to specify which plugins to install. To be able to specify which plugins
to install, you must use the form installer interface.